rename event-card to event-card-mapper
