+++ src/widgets/ui/event-card-mapper/event-card-mapper.vue
