+++ src/widgets/ui/event-card-mapper/index.ts
@@ -1,0 +1,1 @@
+import { default as EventCardMapper } from './event-card-mapper.vue'
