--- src/widgets/ui/event-card/index.ts
@@ -1,3 +1,0 @@
-import EventCard from './event-card.vue'
-
-export { EventCard };
