+++ src/widgets/ui/index.ts
@@ -1,6 +1,6 @@
 export * from './page-header';
 export * from './page-event-header';
-export * from './event-card';
+export * from './event-card-mapper';
 export * from './preview-card-default';
 export * from './page-placeholder';
 export * from './layout-sidebar';
